% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnn.R, R/umap.R
\name{chan_mnn}
\alias{chan_mnn}
\alias{chan_mnn.SingleCellExperiment}
\alias{chan_mnn.default}
\alias{chan_umap.SingleCellExperiment}
\title{Fast mutual nearest neighbors correction with \code{scran.chan}}
\usage{
chan_mnn(object, ...)

\method{chan_mnn}{SingleCellExperiment}(
  object,
  ...,
  dimred = "PCA",
  n_dimred = NULL,
  exprs_values = NULL,
  name = "corrected"
)

\method{chan_mnn}{default}(object, batch = NULL, k = 20L, ..., approximate = TRUE, threads = 1L)

\method{chan_umap}{SingleCellExperiment}(
  object,
  ...,
  exprs_values = "counts",
  dimred = NULL,
  n_dimred = NULL,
  name = "UMAP"
)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{
  Arguments passed on to \code{\link[scran.chan:mnnCorrect.chan]{scran.chan::mnnCorrect.chan}}
  \describe{
    \item{\code{nmads}}{Numeric scalar specifying the number of MADs to use when computing correction vectors.}
    \item{\code{mass.cap}}{Integer scalar specifying the cap on the number of observations to use for center of mass calculations.
A value of 100,000 may be appropriate for speeding up correction of very large datasets.
If this is set to NULL, no cap is used.}
    \item{\code{order}}{Vector containing levels of \code{batch} in the desired merge order.
If \code{NULL}, a suitable merge order is automatically determined.}
    \item{\code{reference.policy}}{String specifying the policy to use to choose the first reference batch.
This can be based on the largest batch (\code{"max-size"}), the most variable batch (\code{"max-variance"}), 
some combination of those two (\code{"max-rss"}) or the first specified input (\code{"input"}).
Only used for automatic merges, i.e., when \code{order=NULL}.}
  }}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{name}{String specifying the name to be used to store the result in the
\link[SingleCellExperiment:reducedDims]{reducedDims} of the output.}

\item{batch}{Vector or factor of length equal to the number of cells,
specifying the batch of origin for each cell. Alternatively NULL if all cells
belong to the same batch.}

\item{k}{Integer scalar specifying the number of neighbors to use when
identifying MNN pairs.}

\item{approximate}{Logical scalar specifying whether to perform an
approximate neighbor search.}

\item{threads}{Integer scalar specifying the number of threads to use.}
}
\description{
Fast mutual nearest neighbors correction with \code{scran.chan}
}
\seealso{
\link[scran.chan:mnnCorrect.chan]{mnnCorrect.chan}
}
