% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{chan_umap}
\alias{chan_umap}
\alias{chan_umap.default}
\title{Fast mutual nearest neighbors correction with \code{scran.chan}}
\usage{
chan_umap(object, ...)

\method{chan_umap}{default}(
  object,
  n_neighbors = 15L,
  min_dist = 0.01,
  n_epochs = -1L,
  ...,
  approximate = TRUE,
  seed = 123456L,
  threads = 1L
)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{Additional arguments passed to default methods.}

\item{n_neighbors}{Integer scalar specifying the number of neighbors to use
in the UMAP algorithm.}

\item{min_dist}{Numeric scalar specifying the minimum distance between
points.}

\item{n_epochs}{Integer scalar specifying the number of epochs to perform. If
set to \code{-1}, an appropriate number of epochs is chosen based on ncol(x).}

\item{approximate}{Logical scalar specifying whether to perform an approximate neighbor search.}

\item{seed}{Integer scalar specifying the seed to use.}

\item{threads}{Integer scalar specifying the number of threads to use.}
}
\description{
Fast mutual nearest neighbors correction with \code{scran.chan}
}
\seealso{
\link[scran.chan:mnnCorrect.chan]{mnnCorrect.chan}
}
