% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSNNGraph.R
\name{clusterSNNGraph}
\alias{clusterSNNGraph}
\alias{clusterSNNGraph.SingleCellExperiment}
\alias{clusterSNNGraph.Seurat}
\alias{clusterSNNGraph.default}
\title{Graph-based clustering with \code{scran.chan}}
\usage{
clusterSNNGraph(object, ...)

\method{clusterSNNGraph}{SingleCellExperiment}(object, ..., dimred = "PCA", n_dimred = NULL, assay = NULL)

\method{clusterSNNGraph}{Seurat}(
  object,
  ...,
  dimred = "PCA",
  n_dimred = NULL,
  assay = NULL,
  layer = NULL
)

\method{clusterSNNGraph}{default}(
  object,
  k = 15L,
  method = "leiden",
  ...,
  scheme = NULL,
  resolution = 1L,
  objective = NULL,
  steps = 4L,
  approximate = TRUE,
  seed = 123456L,
  threads = 1L
)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{Additional arguments passed to default methods.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}

\item{k}{Integer scalar specifying the number of neighbors to use when
identifying MNN pairs.}

\item{method}{String specifying the community detection method to use.
Options are multi-level (\code{"multilevel"}), Walktrap (\code{"walktrap"}) or Leiden
(\code{"leiden"}).}

\item{scheme}{String specifying the weighting scheme to use for constructing
the SNN graph. This can be \code{"rank"} (default), \code{"jaccard"} or \code{"number"}.}

\item{resolution}{Numeric scalar specifying the resolution to use for
multi-level or Leiden clustering.}

\item{objective}{String specifying the objective function to use for Leiden
clustering: "CPM" or "modularity" (default).}

\item{steps}{Integer scalar specifying the number of steps to use for
Walktrap clustering.}

\item{approximate}{Logical scalar specifying whether to perform an
approximate neighbor search.}

\item{seed}{Integer scalar specifying the seed to use for multi-level or
Leiden clustering.}

\item{threads}{Integer scalar specifying the number of threads to use.}
}
\value{
An integer vector with cluster assignments for each cell. Each method
may also return additional attributes.
\itemize{
\item For method=\code{"multilevel"}, we have:
\itemize{
\item \code{levels}, a list of integer vectors with cluster assignments for each
cell at each level. Assignments are sorted by decreasing resolution (i.e.,
fewer, larger clusters).
\item \code{modularity}, a numeric vector containing the modularity of each level.
\item \code{best}, the level with the lowest modularity.
}
\item For method=\code{"leiden"}, we have:
\itemize{
\item \code{quality}: a numeric scalar containing the quality of the clustering
(either the modularity or a related score).
}
\item For method=\code{"walktrap"}, we have:
\itemize{
\item \code{merges}: an integer matrix specifying how the clusters were merged to
obtain membership. Each row corresponds to a merge step and contains the
IDs of the temporary clusters (not the same as those in membership).
\item \code{modularity}: a numeric vector containing the modularity before and
after each merge step.
}
}
}
\description{
Graph-based clustering with \code{scran.chan}
}
\seealso{
\link[scran.chan:clusterSNNGraph.chan]{clusterSNNGraph.chan}
}
