% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\alias{downsample.SingleCellExperiment}
\alias{downsample.default}
\title{Downsample cells based on their neighbors}
\usage{
downsample(object, ...)

\method{downsample}{SingleCellExperiment}(object, ..., dimred = "PCA", n_dimred = NULL, exprs_values = NULL)

\method{downsample}{default}(object, ..., threads = 1L)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{
  Arguments passed on to \code{\link[scran.chan:downsampleByNeighbors.chan]{scran.chan::downsampleByNeighbors.chan}}
  \describe{
    \item{\code{k}}{Integer scalar specifying the number of neighbors to use.
Larger values result in greater downsampling, roughly at a ratio of \code{k/2}.}
    \item{\code{approximate}}{Logical scalar indicating whether an approximate neighbor search should be performed.}
  }}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{threads}{Integer scalar specifying the number of threads to use.}
}
\value{
List containing:
\itemize{
\item \code{chosen}, an integer vector containing the column indices of \code{x} to retain.
\item \code{assigned}, integer vector of length equal to \code{ncol(x)},
containing the column index of the representative cell for each cell in \code{x}.
}
}
\description{
Downsample cells based on their neighbors
}
\seealso{
\link[scran.chan:downsampleByNeighbors.chan]{downsampleByNeighbors.chan}
}
