% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNormAndPCA.R
\name{logNormAndPCA}
\alias{logNormAndPCA}
\alias{logNormAndPCA.SingleCellExperiment}
\alias{logNormAndPCA.Seurat}
\alias{logNormAndPCA.default}
\title{Perform Principal components analysis with \code{scran.chan}}
\usage{
logNormAndPCA(object, ...)

\method{logNormAndPCA}{SingleCellExperiment}(object, ..., name = "PCA")

\method{logNormAndPCA}{Seurat}(object, ..., assay = NULL, layer = "counts", name = "PCA")

\method{logNormAndPCA}{default}(
  object,
  d = 50L,
  scale = FALSE,
  ...,
  subset_row = NULL,
  batch = NULL,
  batch_method = NULL,
  threads = NULL
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: additional argumentds passed to \link{logNormCounts}.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
default methods.
}}

\item{name}{String specifying the name to be used to store the result in the
\link[SingleCellExperiment:reducedDims]{reducedDims} or
\link[SeuratObject:Seurat-class]{reductions} of the output.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}

\item{d}{Integer scalar specifying the number of top PCs to obtain.}

\item{scale}{Logical scalar indicating whether to scale rows to unit
variance.}

\item{subset_row}{Integer, logical or character vector specifying which
features to use in the PCA (e.g., highly variable genes). If NULL, all
features in x are used.}

\item{batch}{Vector or factor of length equal to the number of cells,
specifying the batch of origin for each cell. Alternatively NULL if all cells
belong to the same batch.}

\item{batch_method}{String indicating how batch should be handled (if it is
supplied). \code{"block"} is equivalent to linear regression on x prior to PCA,
while \code{"weight"} will only weight each batch so that they contribute equally
to the PCA.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\link[parallel:detectCores]{detectCores}.}
}
\value{
\itemize{
\item \code{default} method: A numeric matrix where rows are cells and columns are
the two dimensions of the embedding.
\item \code{SingleCellExperiment} method: embedding was added into
\link[SingleCellExperiment:reducedDims]{reducedDims} named as \code{name}.
\item \code{Seurat} method: embedding was added into
\link[SeuratObject:Seurat-class]{reductions} named as \code{name}.
}
}
\description{
Perform Principal components analysis with \code{scran.chan}
}
\seealso{
\itemize{
\item \link{logNormCounts}
\item \link{runPCA}
}
}
