% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNormCounts.R
\name{logNormCounts}
\alias{logNormCounts}
\alias{logNormCounts.SingleCellExperiment}
\alias{logNormCounts.Seurat}
\alias{logNormCounts.default}
\title{Log-normalize the counts}
\usage{
logNormCounts(object, ...)

\method{logNormCounts}{SingleCellExperiment}(object, size_factors = NULL, ..., assay = "counts")

\method{logNormCounts}{Seurat}(object, ..., assay = NULL, layer = "counts")

\method{logNormCounts}{default}(
  object,
  size_factors = NULL,
  batch = NULL,
  batch_mode = NULL,
  ...,
  force_integer = TRUE,
  no_sparse_copy = TRUE,
  threads = NULL
)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: cannot be used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} methods.
}}

\item{size_factors}{A numeric vector of length equal to the number of cells
in x, containing positive size factors for all cells.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}

\item{batch}{Vector or factor of length equal to the number of cells,
specifying the batch of origin for each cell. Alternatively NULL if all cells
belong to the same batch.}

\item{batch_mode}{String indicating how batch should be handled when
centering the size factors. If \code{"lowest"}, we downscale all batches to the
coverage of the lowest batch. If \code{"perblock"}, we scale each batch to a mean
of 1. Default: \code{"perblock"}.}

\item{force_integer}{Logical scalar indicating whether double-precision
\code{object} should be forced into integers.}

\item{no_sparse_copy}{Logical scalar indicating whether we should avoid a
copy when object is a \link[Matrix:dgCMatrix-class]{dgCMatrix} This is more
memory efficient if the data has already been loaded into memory. If \code{TRUE},
any setting of force.integer is ignored.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\link[parallel:detectCores]{detectCores}.}
}
\value{
A list like \code{x} where the count matrix is replaced with a log-expression matrix.
An additional numeric \code{size.factors} vector is also present containing the (centered) size factors.
}
\description{
Log-normalize the counts
}
\seealso{
\link[scran.chan:logNormCounts.chan]{logNormCounts.chan}
}
