% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickMNN.R
\name{quickMNN}
\alias{quickMNN}
\alias{quickMNN.SingleCellExperiment}
\alias{quickMNN.Seurat}
\title{Quick fastMNN with \code{scran.chan}}
\usage{
quickMNN(object, ...)

\method{quickMNN}{SingleCellExperiment}(
  object,
  batch,
  ...,
  k = 15L,
  nmads = 3L,
  mass_cap = NULL,
  order = NULL,
  reference_policy = NULL,
  approximate = TRUE,
  threads = NULL,
  name = "corrected"
)

\method{quickMNN}{Seurat}(
  object,
  batch,
  ...,
  k = 15L,
  nmads = 3L,
  mass_cap = NULL,
  order = NULL,
  reference_policy = NULL,
  approximate = TRUE,
  threads = NULL,
  name = "corrected"
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{
  Arguments passed on to \code{\link[=runPCA]{runPCA}}
  \describe{
    \item{\code{}}{}
  }}

\item{batch}{Vector or factor of length equal to the number of cells,
specifying the batch of origin for each cell. Alternatively NULL if all cells
belong to the same batch.}

\item{k}{Integer scalar specifying the number of neighbors to use when
identifying MNN pairs.}

\item{nmads}{Numeric scalar specifying the number of MADs to use when computing correction vectors.}

\item{mass_cap}{Integer scalar specifying the cap on the number of
observations to use for center of mass calculations. A value of 100,000 may
be appropriate for speeding up correction of very large datasets. If this is
set to \code{NULL}, no cap is used.}

\item{order}{Vector containing levels of \code{batch} in the desired merge order.
If \code{NULL}, a suitable merge order is automatically determined.}

\item{reference_policy}{String specifying the policy to use to choose the
first reference batch. This can be based on the largest batch (\code{"max-size"}),
the most variable batch (\code{"max-variance"}), some combination of those two
(\code{"max-rss"}) or the first specified input ("input"). Only used for automatic
merges, i.e., when order=\code{NULL}.}

\item{approximate}{Logical scalar specifying whether to perform an
approximate neighbor search.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\link[parallel:detectCores]{detectCores}.}

\item{name}{String specifying the name to be used to store the result in the
\link[SingleCellExperiment:reducedDims]{reducedDims} or
\link[SeuratObject:Seurat-class]{reductions} of the output.}
}
\description{
Quick fastMNN with \code{scran.chan}
}
\seealso{
\itemize{
\item \link[scran.chan:logNormCounts.chan]{logNormCounts.chan}
\item \link[scran.chan:runPCA.chan]{runPCA.chan}
\item \link[scran.chan:mnnCorrect.chan]{mnnCorrect.chan}
}
}
