% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCA.R
\name{runPCA}
\alias{runPCA}
\alias{runPCA.SingleCellExperiment}
\alias{runPCA.Seurat}
\alias{runPCA.default}
\title{Perform Principal components analysis with \code{scran.chan}}
\usage{
runPCA(object, ...)

\method{runPCA}{SingleCellExperiment}(
  object,
  ...,
  assay = "logcounts",
  dimred = NULL,
  n_dimred = NULL,
  name = "PCA"
)

\method{runPCA}{Seurat}(
  object,
  ...,
  assay = NULL,
  layer = "data",
  dimred = NULL,
  n_dimred = NULL,
  name = "PCA"
)

\method{runPCA}{default}(
  object,
  d = 50L,
  scale = FALSE,
  ...,
  subset_row = NULL,
  batch = NULL,
  batch_method = NULL,
  force_integer = TRUE,
  no_sparse_copy = TRUE,
  threads = NULL
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: cannot be used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
default methods.
}}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{name}{String specifying the name to be used to store the result in the
\link[SingleCellExperiment:reducedDims]{reducedDims} or
\link[SeuratObject:Seurat-class]{reductions} of the output.}

\item{layer}{Name of the layer to get from the assay data.}

\item{d}{Integer scalar specifying the number of top PCs to obtain.}

\item{scale}{Logical scalar indicating whether to scale rows to unit
variance.}

\item{subset_row}{Integer, logical or character vector specifying which
features to use in the PCA (e.g., highly variable genes). If NULL, all
features in x are used.}

\item{batch}{Vector or factor of length equal to the number of cells,
specifying the batch of origin for each cell. Alternatively NULL if all cells
belong to the same batch.}

\item{batch_method}{String indicating how batch should be handled (if it is
supplied). \code{"block"} is equivalent to linear regression on x prior to PCA,
while \code{"weight"} will only weight each batch so that they contribute equally
to the PCA.}

\item{force_integer}{Logical scalar indicating whether double-precision x
should be forced into integers.}

\item{no_sparse_copy}{Logical scalar indicating whether we should avoid a
copy when object is a \link[Matrix:dgCMatrix-class]{dgCMatrix} This is more
memory efficient if the data has already been loaded into memory. If \code{TRUE},
any setting of force.integer is ignored.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\link[parallel:detectCores]{detectCores}.}
}
\value{
\itemize{
\item \code{default} method: A numeric matrix where rows are cells and columns are
the two dimensions of the embedding.
\item \code{SingleCellExperiment} method: embedding was added into
\link[SingleCellExperiment:reducedDims]{reducedDims} named as \code{name}.
\item \code{Seurat} method: embedding was added into
\link[SeuratObject:Seurat-class]{reductions} named as \code{name}.
}
}
\description{
Perform Principal components analysis with \code{scran.chan}
}
\seealso{
\link[scran.chan:runPCA.chan]{runPCA.chan}
}
