% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUMAP.R
\name{runUMAP}
\alias{runUMAP}
\alias{runUMAP.SingleCellExperiment}
\alias{runUMAP.default}
\title{Compute the uniform manifold approximation and projection with \code{scran.chan}}
\usage{
runUMAP(object, ...)

\method{runUMAP}{SingleCellExperiment}(
  object,
  ...,
  dimred = "PCA",
  n_dimred = NULL,
  exprs_values = NULL,
  name = "UMAP"
)

\method{runUMAP}{default}(
  object,
  n_neighbors = 15L,
  min_dist = 0.01,
  n_epochs = -1L,
  ...,
  approximate = TRUE,
  seed = 123456L,
  threads = 1L
)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{Additional arguments passed to default methods.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{name}{String specifying the name to be used to store the result in the
\link[SingleCellExperiment:reducedDims]{reducedDims} of the output.}

\item{n_neighbors}{Integer scalar specifying the number of neighbors to use
in the UMAP algorithm.}

\item{min_dist}{Numeric scalar specifying the minimum distance between
points.}

\item{n_epochs}{Integer scalar specifying the number of epochs to perform. If
set to \code{-1}, an appropriate number of epochs is chosen based on ncol(x).}

\item{approximate}{Logical scalar specifying whether to perform an approximate neighbor search.}

\item{seed}{Integer scalar specifying the seed to use.}

\item{threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A numeric matrix where rows are cells and columns are the two
dimensions of the embedding.
}
\description{
Compute the uniform manifold approximation and projection with \code{scran.chan}
}
\seealso{
\link[scran.chan:runUMAP.chan]{runUMAP.chan}
}
