% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMarkers.R
\name{scoreMarkers}
\alias{scoreMarkers}
\alias{scoreMarkers.SingleCellExperiment}
\alias{scoreMarkers.Seurat}
\alias{scoreMarkers.default}
\title{Score marker genes}
\usage{
scoreMarkers(object, ...)

\method{scoreMarkers}{SingleCellExperiment}(object, ..., assay = "logcounts")

\method{scoreMarkers}{Seurat}(object, ..., assay = NULL, layer = "data")

\method{scoreMarkers}{default}(
  object,
  groups,
  lfc = 0L,
  simple_means_only = TRUE,
  sort_by = "cohen.rank",
  all_pairwise = FALSE,
  ...,
  batch = NULL,
  force_integer = FALSE,
  no_sparse_copy = TRUE,
  threads = NULL
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: cannot be used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} methods.
}}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}

\item{groups}{A vector specifying the group assignment for each cell in x.}

\item{lfc}{Non-negative numeric scalar specifying the log-fold change threshold to use.}

\item{simple_means_only}{Logical scalar indicating whether to only report the
means for the simple effect sizes, i.e., log-fold change and delta-detected.}

\item{sort_by}{String specifying the column to use for sorting genes in
descending order (except if it ends with \code{.rank}, in which case it is sorted
in ascending order). If \code{NULL}, no sorting is performed.}

\item{all_pairwise}{Logical scalar indicating whether to report the full
effects for every pairwise comparison.}

\item{batch}{Vector or factor of length equal to the number of cells,
specifying the batch of origin for each cell. Alternatively \code{NULL} if all
cells belong to the same batch.}

\item{force_integer}{Logical scalar indicating whether double-precision
\code{object} should be forced into integers.}

\item{no_sparse_copy}{Logical scalar indicating whether we should avoid a
copy when object is a \link[Matrix:dgCMatrix-class]{dgCMatrix} This is more
memory efficient if the data has already been loaded into memory. If \code{TRUE},
any setting of force.integer is ignored.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\link[parallel:detectCores]{detectCores}.}
}
\value{
A list of data frame of marker statistics.
Each data frame corresponds to a group in \code{groups} and contains:
\itemize{
\item \code{mean}: the mean expression across all cells in the current group.
\item \code{detected}: proportion of cells with detectable expression in the current
group.
\item \code{logFC}: the mean of the log-fold changes in expression compared to other
groups.
\item \code{delta.detected}: the mean of the difference in the detected proportions
compared to other groups.
\item \code{cohen.min}: the smallest Cohen's d across all pairwise comparisons
involving the current group.
\item \code{cohen.mean}: the mean Cohen's d across all pairwise comparisons involving
the current group.
\item \code{cohen.rank}: the minimum rank of the Cohen's d across all pairwise
comparisons.
\item \code{auc.min}: the smallest AUC across all pairwise comparisons involving the
current group.
\item \code{auc.mean}: the mean AUC across all pairwise comparisons involving the
current group.
\item \code{auc.rank}: the minimum rank of the AUC across all pairwise comparisons.
}

Rows are sorted by the specified column in \code{sort.by}.

If \code{simple_means_only=FALSE}, \code{logFC} and \code{delta.detected} are
renamed to \code{logFC.mean} and \code{delta.detected.mean}, respectively.
In addition, the corresponding \verb{*.min} and \verb{*.rank} columns are
also reported.
These can be interpreted in a similar manner as \code{cohen.min},
\code{cohen.rank}, etc. for these effect sizes.

If \code{batch} is supplied, this list will also contain \code{per.batch}.
This is a list containing \code{mean} and \code{detected}, each of which are
lists of data frames containing the batch-specific statistics for each group.
(In this case, \code{statistics} contains the averaged statistics for each
gene across batches)

If \code{all_pairwise=TRUE}, this list will also contain \code{pairwise}, a list of
lists of data frames. Each data frame contains the statistics for the
pairwise comparison between groups, e.g., \code{pairwise$A$B} contains the
statistics for \verb{A versus B} where large effects correspond to upregulation in
A. Note that rows correspond to the order in \code{object}. \code{sort_by} has no
effect on these data frames.
}
\description{
Score marker genes
}
\details{
\code{min} is the most stringent summary statistic for identifying upregulated genes, 
as a gene must be strongly upregulated in every pairwise comparison to achieve a large \code{min} effect.
\code{rank} is the most generous as a gene may be highly ranked in any one pairwise comparison to achieve a high \code{rank},
and is useful for identifying the combination of best genes that distinguish a cluster from the others.
\code{mean} lies in between these two extremes.

Cohen's d and the AUC are generally give quite similar rankings,
but both are reported for some verisimilitude.
}
\seealso{
\link[scran.chan:scoreMarkers.chan]{scoreMarkers.chan}
}
