% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{chan_umap}
\alias{chan_umap}
\alias{chan_umap.default}
\title{Compute the uniform manifold approximation and projection with \code{scran.chan}}
\usage{
chan_umap(object, ...)

\method{chan_umap}{default}(
  object,
  n_neighbors = 15L,
  min_dist = 0.01,
  n_epochs = -1L,
  ...,
  approximate = TRUE,
  seed = 123456L,
  threads = 1L
)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{Additional arguments passed to default methods.}

\item{n_neighbors}{Integer scalar specifying the number of neighbors to use
in the UMAP algorithm.}

\item{min_dist}{Numeric scalar specifying the minimum distance between
points.}

\item{n_epochs}{Integer scalar specifying the number of epochs to perform. If
set to \code{-1}, an appropriate number of epochs is chosen based on ncol(x).}

\item{approximate}{Logical scalar specifying whether to perform an approximate neighbor search.}

\item{seed}{Integer scalar specifying the seed to use.}

\item{threads}{Integer scalar specifying the number of threads to use.}
}
\value{
By default, a numeric matrix where rows are cells and columns are the two dimensions of the embedding.

If any of the parameters are vectors or \code{drop = FALSE}, we assume that the user is requesting a parameter sweep.
A list is returned containing \code{parameters}, a data.frame with each relevant combination of parameters;
and \code{results}, a list of numeric matrices of length equal to the number of rows of \code{parameters}, containing the embeddings for each combination.
}
\description{
Compute the uniform manifold approximation and projection with \code{scran.chan}
}
\seealso{
\link[scran.chan:mnnCorrect.chan]{mnnCorrect.chan}
}
