% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\alias{downsample.SingleCellExperiment}
\alias{downsample.default}
\title{Downsample cells based on their neighbors}
\usage{
downsample(object, ...)

\method{downsample}{SingleCellExperiment}(
  object,
  ...,
  dimred = "PCA",
  n_dimred = NULL,
  assay = NULL,
  layer = NULL
)

\method{downsample}{SingleCellExperiment}(
  object,
  ...,
  dimred = "PCA",
  n_dimred = NULL,
  assay = NULL,
  layer = NULL
)

\method{downsample}{default}(object, ..., threads = NULL)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{
  Arguments passed on to \code{\link[scran.chan:downsampleByNeighbors.chan]{scran.chan::downsampleByNeighbors.chan}}
  \describe{
    \item{\code{k}}{Integer scalar specifying the number of neighbors to use.
Larger values result in greater downsampling, roughly at a ratio of \code{k/2}.}
    \item{\code{approximate}}{Logical scalar indicating whether an approximate neighbor search should be performed.}
  }}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\link[parallel:detectCores]{detectCores}.}
}
\value{
List containing:
\itemize{
\item \code{chosen}, an integer vector containing the column indices of \code{x} to retain.
\item \code{assigned}, integer vector of length equal to \code{ncol(x)},
containing the column index of the representative cell for each cell in \code{x}.
}
}
\description{
Downsample cells based on their neighbors
}
\seealso{
\link[scran.chan:downsampleByNeighbors.chan]{downsampleByNeighbors.chan}
}
