% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCA.R
\name{runPCA}
\alias{runPCA}
\alias{runPCA.SingleCellExperiment}
\alias{runPCA.default}
\title{Perform Principal components analysis with \code{scran.chan}}
\usage{
runPCA(object, ...)

\method{runPCA}{SingleCellExperiment}(
  object,
  ...,
  exprs_values = "counts",
  dimred = NULL,
  n_dimred = NULL,
  size_factors = NULL,
  name = "PCA"
)

\method{runPCA}{default}(
  object,
  d = 50L,
  scale = FALSE,
  ...,
  size_factors = NULL,
  subset_row = NULL,
  batch = NULL,
  norm_batch = NULL,
  pca_batch = NULL,
  force_integer = TRUE,
  no_sparse_copy = TRUE,
  threads = 1L
)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{Additional arguments passed to default methods.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{size_factors}{A numeric vector of length equal to the number of cells
in x, containing positive size factors for all cells.}

\item{name}{String specifying the name to be used to store the result in the
\link[SingleCellExperiment:reducedDims]{reducedDims} of the output.}

\item{d}{Integer scalar specifying the number of top PCs to obtain.}

\item{scale}{Logical scalar indicating whether to scale rows to unit
variance.}

\item{subset_row}{Integer, logical or character vector specifying which
features to use in the PCA (e.g., highly variable genes). If NULL, all
features in x are used.}

\item{batch}{Vector or factor of length equal to the number of cells,
specifying the batch of origin for each cell. Alternatively NULL if all cells
belong to the same batch.}

\item{norm_batch}{String indicating how batch should be handled when
centering the size factors. If \code{"lowest"}, we downscale all batches to the
coverage of the lowest batch. If \code{"perblock"}, we scale each batch to a mean
of 1.}

\item{pca_batch}{String indicating how batch should be handled (if it is
supplied). "block" is equivalent to linear regression on x prior to PCA,
while "weight" will only weight each batch so that they contribute equally to
the PCA.}

\item{force_integer}{Logical scalar indicating whether double-precision x
should be forced into integers.}

\item{no_sparse_copy}{Logical scalar indicating whether we should avoid a
copy when object is a \link[Matrix:dgCMatrix-class]{dgCMatrix} This is more
memory efficient if the data has already been loaded into memory. If TRUE,
any setting of force.integer is ignored.}

\item{threads}{Integer scalar specifying the number of threads to use.}
}
\value{
A numeric matrix where rows are cells and columns are the two
dimensions of the embedding.
}
\description{
Perform Principal components analysis with \code{scran.chan}
}
\seealso{
\link[scran.chan:runPCA.chan]{runPCA.chan}
}
