% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreFeatureSet.R
\name{scoreFeatureSet}
\alias{scoreFeatureSet}
\alias{scoreFeatureSet.SingleCellExperiment}
\alias{scoreFeatureSet.Seurat}
\alias{scoreFeatureSet.default}
\title{Score feature set activity for each cell}
\usage{
scoreFeatureSet(object, ...)

\method{scoreFeatureSet}{SingleCellExperiment}(object, ..., assay = "logcounts")

\method{scoreFeatureSet}{Seurat}(object, ..., assay = NULL, layer = "data")

\method{scoreFeatureSet}{default}(
  object,
  features,
  batch = NULL,
  scale = FALSE,
  ...,
  force_integer = FALSE,
  no_sparse_copy = TRUE,
  threads = NULL
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: cannot be used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} methods.
}}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}

\item{features}{Integer, logical or character vector specifying the features that belong to the set.}

\item{batch}{Vector or factor of length equal to the number of cells,
specifying the batch of origin for each cell. Alternatively \code{NULL} if all
cells belong to the same batch.}

\item{scale}{Logical scalar indicating whether to scale rows to unit
variance.}

\item{force_integer}{Logical scalar indicating whether double-precision
\code{object} should be forced into integers.}

\item{no_sparse_copy}{Logical scalar indicating whether we should avoid a
copy when object is a \link[Matrix:dgCMatrix-class]{dgCMatrix} This is more
memory efficient if the data has already been loaded into memory. If \code{TRUE},
any setting of force.integer is ignored.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\link[parallel:detectCores]{detectCores}.}
}
\value{
List containing \code{scores}, a numeric vector of per-cell scores for each column in \code{x};
and \code{weights}, a numeric vector of per-feature weights for each feature in \code{features} (ordered by their row indices in \code{x}).
}
\description{
Compute per-cell scores for a given feature set, defined as the column sums of a rank-1 approximation to the submatrix for the feature set.
This uses the same approach implemented in the \pkg{GSDecon} package from Jason Hackney.
}
\seealso{
\link[scran.chan:scoreFeatureSet.chan]{scoreFeatureSet.chan}
}
